// ui/app/src/models/apiEndpoints.ts
export enum ApiEndpoint {
  Workspaces = "workspaces",
  WorkspaceServices = "workspace-services",
  UserResources = "user-resources",
  SharedServices = "shared-services",
  Requests = "requests",
  AirlockRequests = "requests",
  AirlockLink = "link",
  AirlockSubmit = "submit",
  AirlockCancel = "cancel",
  AirlockRevoke = "revoke",
  AirlockReview = "review",
  AirlockCreateReviewResource = "review-user-resource",
  WorkspaceRequests = "workspace-requests",
  WorkspaceRequestSubmit = "submit",
  WorkspaceRequestCancel = "cancel",
  WorkspaceRequestReview = "review",
  WorkspaceTemplates = "workspace-templates",
  WorkspaceServiceTemplates = "workspace-service-templates",
  UserResourceTemplates = "user-resource-templates",
  SharedServiceTemplates = "shared-service-templates",
  Operations = "operations",
  History = "history",
  InvokeAction = "invoke-action",
  Costs = "costs",
  Metadata = ".metadata",
  Health = "health",
  Users = "users",
  AssignableUsers = "assignable-users",
  Roles = "roles"
}